+++ src/Data/Private/ReplayFile.cpp
@@ -105,7 +105,7 @@ namespace ScoreSaber::Data::Private
         CutDirection = _CutDirection;
     }
 
-    NoteID::NoteID(float _Time, int _LineLayer, int _LineIndex, int _ColorType, int _CutDirection, int _GameplayType, int _ScoringType, int _CutDirectionAngleOffset)
+    NoteID::NoteID(float _Time, int _LineLayer, int _LineIndex, int _ColorType, int _CutDirection, int _GameplayType, int _ScoringType, float _CutDirectionAngleOffset)
     {
         Time = _Time;
         LineLayer = _LineLayer;
